+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/model/data/BluetoothData.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014. Queen Mary University of London
+ * Copyright (c) 2015. Queen Mary University of London
  * Kleomenis Katevas, k.katevas@qmul.ac.uk
  *
  * This file is part of SensingKit-Android library.
@@ -19,24 +19,19 @@
  * along with SensingKit-Android.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.sensingkit.sensingkitlib.modules;
+package org.sensingkit.sensingkitlib.model.data;
 
-public enum SensorModuleType {
-    ACCELEROMETER,
-    GRAVITY,
-    LINEAR_ACCELERATION,
-    GYROSCOPE,
-    ROTATION,
-    MAGNETOMETER,
-    AMBIENT_TEMPERATURE,
-    STEP_DETECTOR,
-    STEP_COUNTER,
-    LIGHT,
-    LOCATION,
-    ACTIVITY,
-    BATTERY,
-    SCREEN_STATUS,
-    AUDIO_RECORDER,
-    AUDIO_LEVEL,
-    BLUETOOTH
+public class BluetoothData extends AbstractData {
+
+    @SuppressWarnings("unused")
+    private static final String TAG = "BluetoothData";
+
+    public BluetoothData(long timestamp) {
+        super(timestamp);
+    }
+
+    @Override
+    public String getDataInCSV() {
+        return null;
+    }
 }