+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### 0.1.1
+  - Added support for Bluetooth sensor module
   - DataInterface Method getDataInString() has been renamed to getDataInCSV()
 
 ### 0.1.0